requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
asyncpg==0.31.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
colorama==0.4.6
crewai==1.9.3
cryptography==46.0.5
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.129.0
fastapi-cli==0.0.23
fastapi-cloud-cli==0.13.0
fastar==0.8.0
filelock==3.24.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
instructor==1.12.0
Jinja2==3.1.6
jiter==0.10.0
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.10
langchain-core==1.2.13
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.23.3
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
nodeenv==1.10.0
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.1
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
platformdirs==4.9.2
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.1
propcache==0.4.1
protobuf==5.29.6
pyasn1==0.6.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-extra-types==2.11.0
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.11.0
pypdfium2==5.4.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.2
rich-toolkit==0.19.4
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
scalar_fastapi==1.6.2
sentry-sdk==2.53.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
sqlmodel==0.0.34
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.20.3
tomli==2.0.2
tomli_w==1.1.0
tqdm==4.67.3
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uv==0.9.30
uvicorn==0.41.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
groq
python-dotenv